Rename `uint::bingcd` as `::gcd`
